% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{learndrake-package}
\alias{learndrake-package}
\alias{learndrake}
\title{learndrake: materials to learn machine learning workflow management with drake}
\description{
Machine learning workflows can be difficult to manage.
A single round of computation can take several hours to complete,
and routine updates to the code and data tend to invalidate
hard-earned results. You can enhance the maintainability, hygiene,
speed, scale, and reproducibility of such projects with
the drake R package. drake resolves the dependency structure
of your analysis pipeline, skips tasks that are already up to date,
executes the rest with optional distributed computing,
and organizes the output so you rarely have to think about data files.
The slides, notebooks, and Shiny apps in this package
teach how to create and maintain machine
learning projects with drake-powered automation.
}
\examples{
\dontrun{
view_slides()
write_tutorials()
launch_app("flow")
launch_app("plans")
launch_app("drakeplanner")
}
}
\references{
\url{https://github.com/wlandau/learndrake}
}
\author{
William Michael Landau \email{will.landau@gmail.com}
}
