% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{save_app}
\alias{save_app}
\title{Save the files for a supporting Shiny app}
\usage{
save_app(app = c("flow", "plans", "drakeplanner"), path = app,
  overwrite = FALSE)
}
\arguments{
\item{app}{Character, name of the app to launch. Must be one of
\code{"flow"}, \code{"plans"}, or \code{"drakeplanner"}.}

\item{path}{Character, where to write the files}

\item{overwrite}{Logical, whether to over}
}
\description{
The notebooks \code{3-flow.Rmd} and \code{4-plans.Rmd}
(see \code{\link[=save_notebooks]{save_notebooks()}}) come with supporting Shiny apps
to conduct the exercises. Use \code{save_app()} to save the files
for an so you can deploy it to Shiny Server.
}
\details{
For \code{3-flow.Rmd}, \code{save_app("flow")} saves
a collection of exercises for iterating on a \code{drake} workflow.
For \code{4-plans.Rmd}, \code{save_app("plans")} saves guided exercises
on constructing \code{drake} plans,
and \code{save_app("drakeplanner")} saves an app
for visualizing and understanding plans.
}
\examples{
\dontrun{
save_app("flow")
save_app("plans")
save_app("drakeplanner")
}
}
\seealso{
\code{\link[=save_notebooks]{save_notebooks()}}, \code{\link[=view_slides]{view_slides()}}
}
