% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebooks.R
\name{save_notebooks}
\alias{save_notebooks}
\title{Save the tutorials to your computer}
\usage{
save_notebooks(path = "learndrake-notebooks", overwrite = FALSE)
}
\arguments{
\item{path}{Character, where to write the files.}

\item{overwrite}{Logical, whether to overwrite the files.}
}
\description{
The tutorials are a sequence of R notebooks
and supporting files.
}
\examples{
\dontrun{
save_notebooks()
}
}
\seealso{
\code{\link[=view_slides]{view_slides()}}, \code{\link[=launch_app]{launch_app()}}
}
