% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{save_slides}
\alias{save_slides}
\title{Save the introductory slides}
\usage{
save_slides(path = "learndrake-slides", overwrite = FALSE)
}
\arguments{
\item{path}{Character, where to write the files.}

\item{overwrite}{Logical, whether to overwrite the files.}
}
\description{
Save the slides to a folder.
}
\examples{
\dontrun{
save_slides()
}
}
\seealso{
\code{\link[=view_slides]{view_slides()}}, \code{\link[=save_notebooks]{save_notebooks()}}, \code{\link[=launch_app]{launch_app()}}
}
