% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{view_slides}
\alias{view_slides}
\title{View the introductory slides}
\usage{
view_slides(browse = TRUE)
}
\arguments{
\item{browse}{Logical, whether to actually view the slides.
For testing purposes only.}
}
\description{
Open the slides in a web browser.
}
\examples{
\dontrun{
view_slides()
}
}
\seealso{
\code{\link[=save_slides]{save_slides()}}, \code{\link[=save_notebooks]{save_notebooks()}}, \code{\link[=launch_app]{launch_app()}}
}
